Split "settings" views into a separate module.
