+++ perma_web/perma/templates/docs/index.html
@@ -90,7 +90,7 @@ Welcome to the Perma.cc user guide. The best place to learn about how Perma.cc w
       <p class="body-text">We also offer a variety of paid usage plans for other groups and for individuals. To learn more about paid usage plans, please visit the <a href="{% url 'docs_accounts' %}#usage-plans">Accounts and Usage Plans</a> portion of our user guide.</p>
 
       <h3 id="creating-links" class="body-bh">Preserving records</h3>
-      <p class="body-text">To preserve a web page, log in to your account, enter the URL of the page you want to preserve and click the "Create Perma Link" button, or try one of our <a href="{% url 'user_management_settings_tools' %}">Browser Tools</a>. Perma will visit the web page at your direction and create a record of the contents of that page.</p>
+      <p class="body-text">To preserve a web page, log in to your account, enter the URL of the page you want to preserve and click the "Create Perma Link" button, or try one of our <a href="{% url 'settings_tools' %}">Browser Tools</a>. Perma will visit the web page at your direction and create a record of the contents of that page.</p>
 
       <img src="{{ STATIC_URL }}img/docs/create_link.png" class="img-responsive screen-shot" alt="A screenshot of the Create a Perma Link page. A large input field for the target URL and a 'Create Perma Link' button appear prominently at the top of the page." />
