+++ perma_web/perma/templates/docs/perma-link-creation.html
@@ -71,7 +71,7 @@ This section of the user guide covers how to create new archives and organize yo
       <p class="body-test">For example, this Perma Link - <a href="http://perma.cc/9265-T4NB" target="blank">http://perma.cc/9265-T4NB</a> - points to a Perma Record of the White House homepage on November 18, 2015.</p>
 
       <h3 class="body-bh" id="web-page-preservation">Preservation Methods</h3>
-      <p class="body-text">To preserve a web page, login to your account, enter the URL of the page you want to preserve and click the "Create Perma Link" button, or try one of our <a href="{% url 'user_management_settings_tools' %}">Browser Tools</a>.</p>
+      <p class="body-text">To preserve a web page, login to your account, enter the URL of the page you want to preserve and click the "Create Perma Link" button, or try one of our <a href="{% url 'settings_tools' %}">Browser Tools</a>.</p>
       <img src="{{ STATIC_URL }}img/docs/create_link.png" class="img-responsive screen-shot" alt="A screenshot of the Create a Perma Link page. A large input field for the target URL and a 'Create Perma Link' button appear prominently at the top of the page." />
       <p class="body-text">Perma visits the URL, fetches the contents of the page, makes a record of those contents - a Perma Record - and deposits that record into the permanent collection. Perma also gives you a unique URL - a Perma Link - that's associated with the new Perma Record. </p>
       <p class="body-text">When the capture process is done, Perma shows you the Perma Record so you can verify it. </p>
