+++ perma_web/perma/templates/includes/upper_right_menu.html
@@ -53,14 +53,14 @@
           {% endif %}
           <li><h3 class="dropdown-header">Settings</h3>
             <ul class="dropdown-inner-list">
-              <li><a href="{% url 'user_management_settings_profile' %}">Profile</a></li>
-              <li><a href="{% url 'user_management_settings_password' %}">Password</a></li>
+              <li><a href="{% url 'settings_profile' %}">Profile</a></li>
+              <li><a href="{% url 'settings_password' %}">Password</a></li>
               {% if request.user.can_view_usage_plan %}
-                <li><a href="{% url 'user_management_settings_usage_plan' %}">Usage Plan</a></li>
+                <li><a href="{% url 'settings_usage_plan' %}">Usage Plan</a></li>
               {% endif %}
-              <li><a href="{% url 'user_management_settings_tools' %}">Tools</a></li>
+              <li><a href="{% url 'settings_tools' %}">Tools</a></li>
               {% if request.user.is_organization_user or request.user.is_registrar_user or request.user.has_registrar_pending %}
-                <li><a href="{% url 'user_management_settings_affiliations' %}">Your affiliations</a></li>
+                <li><a href="{% url 'settings_affiliations' %}">Your affiliations</a></li>
               {% endif %}
             </ul>
           </li>
