+++ perma_web/perma/templates/settings-layout.html
@@ -18,18 +18,18 @@
           <nav aria-label="Settings">
             <ul class="nav nav-pills nav-stacked">
 
-              <li{% if this_page == 'settings_profile' %} class="active" aria-current="page"{% endif %}><a href="{% url 'user_management_settings_profile' %}"><span>Profile</span></a></li>
+              <li{% if this_page == 'settings_profile' %} class="active" aria-current="page"{% endif %}><a href="{% url 'settings_profile' %}"><span>Profile</span></a></li>
 
-              <li{% if this_page == 'settings_password' %} class="active" aria-current="page"{% endif %}><a href="{% url 'user_management_settings_password' %}"><span>Password</span></a></li>
+              <li{% if this_page == 'settings_password' %} class="active" aria-current="page"{% endif %}><a href="{% url 'settings_password' %}"><span>Password</span></a></li>
 
               {% if request.user.can_view_usage_plan %}
-              <li{% if this_page == 'settings_usage_plan' %} class="active" aria-current="page"{% endif %}><a href="{% url 'user_management_settings_usage_plan' %}" id="dashboard-settings"><span>Usage Plan</span></a></li>
+              <li{% if this_page == 'settings_usage_plan' %} class="active" aria-current="page"{% endif %}><a href="{% url 'settings_usage_plan' %}" id="dashboard-settings"><span>Usage Plan</span></a></li>
               {% endif %}
 
-              <li{% if this_page == 'settings_tools' %} class="active" aria-current="page"{% endif %}><a href="{% url 'user_management_settings_tools' %}"><span>Tools</span></a></li>
+              <li{% if this_page == 'settings_tools' %} class="active" aria-current="page"{% endif %}><a href="{% url 'settings_tools' %}"><span>Tools</span></a></li>
 
               {% if request.user.is_organization_user or request.user.is_registrar_user or request.user.has_registrar_pending %}
-              <li{% if this_page == 'settings_affiliations' %} class="active" aria-current="page"{% endif %}><a href="{% url 'user_management_settings_affiliations' %}" id="dashboard-settings"><span>Your Affiliations</span></a></li>
+              <li{% if this_page == 'settings_affiliations' %} class="active" aria-current="page"{% endif %}><a href="{% url 'settings_affiliations' %}" id="dashboard-settings"><span>Your Affiliations</span></a></li>
               {% endif %}
 
             </ul><!--/#dashboard-tabs-->
