+++ perma_web/perma/templates/settings/settings-organizations-change-privacy.html
@@ -9,11 +9,11 @@
     {% csrf_token %}
     {% if org.default_to_private %}
       <p class="page-dek">Your Perma Links are currently <strong>Private</strong> by default.</p>
-      <a class="btn cancel" href="{% url 'user_management_settings_affiliations' %}">Cancel</a>
+      <a class="btn cancel" href="{% url 'settings_affiliations' %}">Cancel</a>
       <button type="submit" class="btn">Set to Public</button>
     {% else %}
       <p class="page-dek">Your Perma Links are currently <strong>Public</strong> by default.</p>
-      <a class="btn cancel" href="{% url 'user_management_settings_affiliations' %}">Cancel</a>
+      <a class="btn cancel" href="{% url 'settings_affiliations' %}">Cancel</a>
       <button type="submit" class="btn">Set to Private</button>
     {% endif %}
   </form>
