+++ perma_web/perma/templates/user_management/create-link.html
@@ -50,18 +50,18 @@
           {% if request.user.is_individual %}
             <p class="links-remaining-message">
               {% if subscription_status == 'problem' %}
-                Your subscription is on hold due to a <a href="{% url 'user_management_settings_usage_plan' %}">problem with your credit card</a>.<br>
+                Your subscription is on hold due to a <a href="{% url 'settings_usage_plan' %}">problem with your credit card</a>.<br>
               {% endif %}
               {% if links_remaining == 'Infinity' %}
                 {% if request.user.nonpaying %}
                   Your account has been upgraded to unlimited personal Perma Links.
                 {% else %}
-                  Your <a href="{% url 'user_management_settings_usage_plan' %}">subscription</a> includes the creation of unlimited Perma Links.
+                  Your <a href="{% url 'settings_usage_plan' %}">subscription</a> includes the creation of unlimited Perma Links.
                 {% endif %}
               {% else %}
                   You have <span class="links-remaining">{{ links_remaining }}</span> remaining Perma Links{% if links_remaining_period == "monthly" %} this month{% elif links_remaining_period == "annually" %} this year{% endif %}.
                   {% if not request.user.nonpaying and not subscription_status == 'problem' %}
-                    <br><a href="{% url 'user_management_settings_usage_plan' %}">View your subscription details or get more Perma Links</a>
+                    <br><a href="{% url 'settings_usage_plan' %}">View your subscription details or get more Perma Links</a>
                   {% endif %}
               {% endif %}
             </p>
@@ -77,16 +77,16 @@
             </div>
             <p id="personal-links-banner" class="hide">
               {% if subscription_status == 'problem' %}
-                <a href="{% url 'user_management_settings_usage_plan' %}">Your account is on hold</a>.<br>
+                <a href="{% url 'settings_usage_plan' %}">Your account is on hold</a>.<br>
               {% elif links_remaining != 'Infinity' and not request.user.nonpaying %}
-                get <a href="{% url 'user_management_settings_usage_plan' %}">more personal links</a>
+                get <a href="{% url 'settings_usage_plan' %}">more personal links</a>
                {% endif %}
             </p>
           {% endif %}
         </fieldset>
         {% if suppress_reminder != 'true' %}
           <p id="browser-tools-message" class="limit-{% if request.user.is_individual %}true{% else %}false{% endif %}">
-            To make Perma links more quickly, try our <a href="{% url 'user_management_settings_tools' %}">browser tools</a>.
+            To make Perma links more quickly, try our <a href="{% url 'settings_tools' %}">browser tools</a>.
             <button type="button" class="close-browser-tools btn-link">
               <span aria-hidden="true">&times;</span>
               <span class="sr-only">Close</span>
