+++ perma_web/perma/templates/user_management/manage_single_organization.html
@@ -10,10 +10,10 @@
       <div class="settings-block">
         {% if target_org.default_to_private %}
           <p>Links are <strong>Private</strong> by default</p>
-          <a href="{% url 'user_management_settings_organizations_change_privacy' target_org.id %}" class="btn btn-default btn-xs leave-org-btn" role="button">Set to public</a>
+          <a href="{% url 'settings_organizations_change_privacy' target_org.id %}" class="btn btn-default btn-xs leave-org-btn" role="button">Set to public</a>
         {% else %}
           <p>Links are <strong>Public</strong> by default</p>
-          <a href="{% url 'user_management_settings_organizations_change_privacy' target_org.id %}" class="btn btn-default btn-xs leave-org-btn" role="button">Set to private</a>
+          <a href="{% url 'settings_organizations_change_privacy' target_org.id %}" class="btn btn-default btn-xs leave-org-btn" role="button">Set to private</a>
         {% endif %}
       </div>
       <form method="post" action="">
